clean up code structure a bit
