+++ shared/babylon/types/Interactable.ts
@@ -1,5 +1,9 @@
 import * as BABYLON from 'babylonjs'
 
+export interface InteractableMap {
+  [key: string]: Interactable
+}
+
 export default interface Interactable {
   aggregate: BABYLON.PhysicsAggregate
   name: string
